Pulling flash methods up to HttpSupport.java from AppController.java
