+++ activeweb/src/main/java/org/javalite/activeweb/HttpSupport.java
@@ -112,6 +112,44 @@ public class HttpSupport {
         view(map(values));
     }
 
+    /**
+     * Convenience method, takes in a map of values to flash.
+     *
+     * @see #flash(String, Object)
+     *
+     * @param values values to flash.
+     */
+    protected void flash(Map values){
+        for(Object key:values.keySet() ){
+            flash(key.toString(), values.get(key));
+        }
+    }
+
+    /**
+     * Convenience method, takes in a vararg of values to flash.
+     * Number of values must be even.
+     *
+     * @see #flash(String, Object)
+     * @param values values to flash.
+     */
+    protected void flash(Object ... values){
+        flash(map(values));
+    }
+
+    /**
+     * Sends value to flash. Flash survives one more request.  Using flash is typical
+     * for POST/GET pattern,
+     *
+     * @param name name of value to flash
+     * @param value value to live for one more request in curent session.
+     */
+    protected void flash(String name, Object value) {
+        if (session().get("flasher") == null) {
+            session().put("flasher", new HashMap());
+        }
+        ((Map) session().get("flasher")).put(name, value);
+    }
+
     public class HttpBuilder {
         private ControllerResponse controllerResponse;
         private HttpBuilder(ControllerResponse controllerResponse){
